Move lexer to module
